✨ made lang permissions apply to PATCH

+++ packages/api/src/lang/lang.controller.ts
@@ -8,40 +8,30 @@ import {
   Patch,
   Req,
   UseGuards,
+  UnauthorizedException,
 } from '@nestjs/common'
 import { Request } from 'express'
 import { CreateLangDto } from './dto/create-lang.dto'
 import { UpdateLangDto } from './dto/update-lang.dto'
-import { Lang } from './entities/lang.entity'
 import { LangService } from './lang.service'
 import { getLimitAndCursor } from '../paginator'
 import { JwtAuthGuard } from '../auth/jwt-auth.guard'
 import { UserService } from '../user/user.service'
-import { HanasRequest } from '../auth/checkUser'
-import { classToClass } from 'class-transformer'
-
-const convertPermissionsToOwners = (lang: Lang) => {
-  const langWithOwners = {
-    ...lang,
-    owners: lang.permissions.map((perm) => perm.user),
-  }
-  delete langWithOwners.permissions
-  return langWithOwners
-}
+import { getReqUser } from '../auth/checkUser'
+import { getUserPermissions } from '../lang-permissions/lang-permissions.service'
 
 @Controller()
 export class LangController {
   constructor(
     private readonly langService: LangService,
-    private userService: UserService,
+    private readonly userService: UserService,
   ) {}
 
   @UseGuards(JwtAuthGuard)
   @Post()
   async create(@Body() createLangDto: CreateLangDto, @Req() req: Request) {
-    const user = await this.userService.findOne(
-      (req as HanasRequest).user.username,
-    )
+    const { username } = getReqUser(req)
+    const user = await this.userService.findOne(username)
     return this.langService.create(createLangDto, user)
   }
 
@@ -49,13 +39,6 @@ export class LangController {
   async findAll(@Req() req: Request) {
     const { limit, cursor } = getLimitAndCursor(req)
     return this.langService.findAll(limit, cursor)
-    /*  const convertedLangs = ((langs.data as unknown) as Lang[]).map(
-      convertPermissionsToOwners,
-    )
-    return {
-      data: classToClass(convertedLangs),
-      cursor: langs.cursor,
-    } */
   }
 
   @Get(':id')
@@ -63,13 +46,37 @@ export class LangController {
     return this.langService.findOne(id)
   }
 
+  @UseGuards(JwtAuthGuard)
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updateLangDto: UpdateLangDto) {
+  async update(
+    @Param('id') id: string,
+    @Body() updateLangDto: UpdateLangDto,
+    @Req() req: Request,
+  ) {
+    const { username } = getReqUser(req)
+    const lang = await this.findOne(id)
+    const perms = getUserPermissions(lang, username)
+    if (!perms.changeInfo && (updateLangDto.description || updateLangDto.name))
+      throw new UnauthorizedException(
+        'User does not have permissions to change language information.',
+      )
+    if (!perms.changeId && updateLangDto.id)
+      throw new UnauthorizedException(
+        'User does not have permission to change language id.',
+      )
     return this.langService.update(id, updateLangDto)
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete(':id')
-  remove(@Param('id') id: string): Promise<void> {
+  async remove(@Param('id') id: string, @Req() req: Request): Promise<void> {
+    const { username } = getReqUser(req)
+    const lang = await this.findOne(id)
+    if (lang.owner.username !== username) {
+      throw new UnauthorizedException(
+        'Cannot delete language if user is not owner.',
+      )
+    }
     return this.langService.remove(id)
   }
 }
